requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
pyswisseph==2.10.3.2
pydantic==2.6.0
python-multipart
